fix: rename defaultDocument.tsx -> DefaultDocument.tsx
